Separated portlib imto Port definition and identification modules
